requirements.txt
easyocr
streamlit 
opencv-python 
numpy 
matplotlib 
pillow
ultralytics
python-Levenshtein 
pandas